fixed cookie issue

+++ frontend/TodoFrontend/src/components/TodoView.jsx
@@ -1,10 +1,14 @@
 import React, { useState } from "react";
 import CardList from "./CardList";
 import { useDate } from "../context/DateContext";
-import CreateTodo from "./CreateTodo"; 
+import CreateTodo from "./CreateTodo";
+import Cookies from "js-cookie"; 
+import { getCSRFTokenFromCookie } from "./getcsrf";
+
+
 const TodoView = () => {
   const { selectedDate, setSelectedDate } = useDate();
-  const [isCreateTodoOpen, setIsCreateTodoOpen] = useState(false); 
+  const [isCreateTodoOpen, setIsCreateTodoOpen] = useState(false);
   const [todos, setTodos] = useState([  
     {
       label: "Work",
@@ -24,24 +28,28 @@ const TodoView = () => {
     setIsCreateTodoOpen(true);
   };
 
-  
   const closeCreateTodoModal = () => {
     setIsCreateTodoOpen(false);
   };
 
-  
   const handleTodoAdded = (newTodo) => {
-    
-    setTodos((prevTodos) => [...prevTodos, newTodo]);  
+    setTodos((prevTodos) => [...prevTodos, newTodo]);
     console.log("New Todo Added:", newTodo);
   };
 
   const handleTodoDeleted = async (todoId) => {
     try {
+      const csrftoken=getCSRFTokenFromCookie();
+      if (!csrftoken) {
+        console.error("Token not found");
+        return;
+      }
+
       const response = await fetch(`http://127.0.0.1:8000/todos/${todoId}/`, {
         method: "DELETE",
         headers: {
-          "Authorization": `Bearer ${localStorage.getItem("token")}`,
+          "Content-Type": "application/json", 
+          'X-CSRFToken': csrftoken,
         },
       });
 
@@ -61,23 +69,22 @@ const TodoView = () => {
       <div className="flex justify-between items-center pt-5 w-full mb-6">
         <p className="text-2xl font-semibold">Tasks</p>
         <button
-          onClick={openCreateTodoModal} 
+          onClick={openCreateTodoModal}
           className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
         >
           Add Task
         </button>
       </div>
 
-      
       {isCreateTodoOpen && (
         <CreateTodo
-          onClose={closeCreateTodoModal} 
-          onTodoAdded={handleTodoAdded} 
+          onClose={closeCreateTodoModal}
+          onTodoAdded={handleTodoAdded}
         />
       )}
 
       <div className="cards pt-6">
-        <CardList todos={todos} onDeleteTodo={handleTodoDeleted} />  
+        <CardList todos={todos} onDeleteTodo={handleTodoDeleted} />
       </div>
     </div>
   );
